rec: Add a small regression test for 'max-cache-bogus-ttl'
